Optimize autocomplete filtering and display
DOM rendering is very expensive for large lists, so limit it to 100
entries.

Also, pushing onto the end of an array is much much faster for long lists
than doing [...array, newElement] for long arrays.

This makes the autocomplete very usable on my computer with 100k entries,
whereas before it would block for 10s or so to show a result.

Signed-off-by: Julius Volz <julius.volz@gmail.com>